requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==46.0.6
distro==1.9.0
ecdsa==0.19.2
fastapi==0.135.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.30.0
passlib==1.7.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sse-starlette==3.3.4
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
